creating contrail base class. for now contrail will directly use backbone framework
radial chart will be implemented simlar to composite-y
