+++ src/js/contrail-charts/charts/ChartView.js
@@ -1,6 +1,6 @@
 define([
   'underscore',
-  'contrail-charts/core/View', // Todo use contrailChartsView instead
+  'contrail-charts/contrail/ContrailView', // Todo use contrailChartsView instead
   'contrail-charts/charts/xy/XYChartView',
   'contrail-charts/handlers/BindingHandler'
 ], function (
