+++ src/js/contrail-charts/charts/xy/XYChartView.js
@@ -6,9 +6,9 @@ define([
   // 'contrail-charts/models/MessageConfigModel',
   // 'contrail-charts/models/NavigationConfigModel',
   // 'contrail-charts/models/TooltipConfigModel',
-  'contrail-charts/base/ContrailChartsDataModel',
+  'contrail-charts/contrail/ContrailChartsDataModel',
   // 'contrail-charts/models/DataProvider',
-  'contrail-charts/core/View',
+  'contrail-charts/contrail/ContrailView',
   // 'contrail-charts/views/CompositeYChartView',
   // 'contrail-charts/views/ControlPanelView',
   // 'contrail-charts/views/MessageView',
