+++ src/js/contrail-charts/components/composite-y-chart/composites/ScatterBubbleChartView.js
@@ -4,8 +4,8 @@
 
 define([
   'jquery', 'underscore', 'd3',
-  'contrail-charts/base/ContrailChartsEvents',
-  'contrail-charts/base/ContrailChartsView'
+  'contrail-charts/contrail/ContrailChartsEvents',
+  'contrail-charts/contrail/ContrailChartsView'
 ], function ($, _, d3, Events, ContrailChartsView) {
   var ScatterBubbleChartView = ContrailChartsView.extend({
     tagName: 'div',
