+++ src/js/contrail-charts/components/composite-y-chart/composites/StackedBarChartView.js
@@ -6,19 +6,21 @@ define([
   'jquery',
   'underscore',
   'd3',
-  'contrail-charts/base/ContrailChartsEvents',
-  'contrail-charts/base/ContrailChartsView'
+  'contrail-charts/contrail/ContrailChartsEvents',
+  'contrail-charts/contrail/ContrailChartsView'
 ], function ($, _, d3, Events, ContrailChartsView) {
-  var BarChartView = ContrailChartsView.extend({
+  var StackedBarChartView = ContrailChartsView.extend({
     tagName: 'div',
     className: 'bar-chart',
-    chartType: 'bar',
+    chartType: 'stackedBar',
     renderOrder: 100,
 
     initialize: function (options) {
       // / The config model
       this.config = options.config
       this.axisName = options.axisName
+
+      // The child's params are reset by parent.
       this.eventObject = _.extend({}, Events)
     },
 
@@ -49,20 +51,24 @@ define([
     /**
     * Called by the parent in order to calculate maximum data extents for all of this child's axis.
     * Assumes the params.activeAccessorData for this child view is filled by the parent with the relevent yAccessors for this child only.
-    * Returns an object with following structure: { y1: [0,10], x: [-10,10] }
+    * Returns an object with following structure: { y1: [0,10], x: [-10,10] } - axisName: axisDomain
     */
     calculateAxisDomains: function () {
       var self = this
       var domains = {}
       domains[self.params.plot.x.axis] = self.model.getRangeFor(self.params.plot.x.accessor)
-      domains[self.axisName] = []
       // The domains calculated here can be overriden in the axis configuration.
       // The overrides are handled by the parent.
       _.each(self.params.activeAccessorData, function (accessor) {
         var domain = self.model.getRangeFor(accessor.accessor)
-        domains[self.axisName] = domains[self.axisName].concat(domain)
+        if (_.has(domains, self.axisName)) {
+          // domains[self.axisName][0] = Math.min( domain[0], domains[self.axisName][0] )
+          domains[self.axisName][1] += domain[1]
+        } else {
+          // domains[self.axisName] = [domain[0], domain[1]]
+          domains[self.axisName] = [0, domain[1]]
+        }
       })
-      domains[self.axisName] = d3.extent(domains[self.axisName])
       self.params.handledAxisNames = _.keys(domains)
       return domains
     },
@@ -74,10 +80,9 @@ define([
     calculateScales: function () {},
 
     /**
-     * Renders an empty chart.
-     * Changes chart dimensions if it already exists.
+     * Called by the parent to allow the child to add some initialization code into the provided entering selection.
      */
-    renderSVG: function () {},
+    renderSVG: function (enteringSelection) {},
 
     renderData: function () {
       var self = this
@@ -87,8 +92,6 @@ define([
 
       // Create a flat data structure
       var flatData = []
-      var j
-      var numOfAccessors = _.keys(self.params.activeAccessorData).length
       var xValues = _.pluck(self.getData(), self.params.plot.x.accessor)
       var xValuesExtent = d3.extent(xValues)
       var xRange = [xScale(xValuesExtent[0]), xScale(xValuesExtent[1])]
@@ -98,31 +101,28 @@ define([
       }
       var bandWidth = (0.95 * ((xRange[1] - xRange[0]) / len) - 1)
       var bandWidthHalf = (bandWidth / 2)
-      var innerBandScale = d3.scaleBand().domain(d3.range(numOfAccessors)).range([0, bandWidth]).paddingInner(0.05).paddingOuter(0.05)
-      var innerBandWidth = (innerBandScale.bandwidth())
       _.each(data, function (d) {
-        j = 0
         var x = d[self.params.plot.x.accessor]
+        var stackedY = yScale.domain()[0]
         _.each(self.params.activeAccessorData, function (accessor) {
           var key = accessor.accessor
-          var y = d[key]
           var obj = {
             id: x + '-' + key,
             className: 'bar bar-' + key,
-            x: xScale(x) - bandWidthHalf + innerBandScale(j),
-            y: yScale(y),
-            h: yScale.range()[0] - yScale(y),
-            w: innerBandWidth,
+            x: xScale(x) - bandWidthHalf,
+            y: yScale(stackedY + d[key]),
+            h: yScale.range()[0] - yScale(d[key]),
+            w: bandWidth,
             color: self.getBarColor(accessor, key),
             accessor: accessor,
             data: d
           }
+          stackedY += d[key]
           flatData.push(obj)
-          j++
         })
       })
       // Render the flat data structure
-      console.log('Rendering data in BarChartView: ', flatData, self.config, self.params, self.getName())
+      console.log('Rendering data in BarChartView: ', flatData, self.params, self.getName())
       var svgBarGroups = self.svgSelection().select('g.drawing-' + self.getName()).selectAll('.bar').data(flatData, function (d) { return d.id })
       svgBarGroups.enter().append('rect')
         .attr('class', function (d) { return d.className })
@@ -131,12 +131,12 @@ define([
         .attr('height', 0)
         .attr('width', function (d) { return d.w })
         .on('mouseover', function (d) {
-          // var pos = $(this).offset() // not working in jquery 3
+          // var pos = $( this ).offset() // not working in jquery 3
           self.eventObject.trigger('mouseover', d.data, d.x, d.y, d.accessor)
           d3.select(this).classed('active', true)
         })
         .on('mouseout', function (d) {
-          // var pos = $(this).offset() // not working in jquery 3
+          // var pos = $( this ).offset() // not working in jquery 3
           self.eventObject.trigger('mouseout', d.data, d.x, d.y)
           d3.select(this).classed('active', false)
         })
@@ -158,5 +158,5 @@ define([
     }
   })
 
-  return BarChartView
+  return StackedBarChartView
 })
