+++ src/js/contrail-charts/components/message/MessageConfigModel.js
@@ -5,7 +5,7 @@
 define([
   'jquery',
   'underscore',
-  'contrail-charts/base/ContrailChartsConfigModel'
+  'contrail-charts/contrail/ContrailChartsConfigModel'
 ], function ($, _, ContrailChartsConfigModel) {
   var MessageConfigModel = ContrailChartsConfigModel.extend({
     defaults: {
