+++ src/js/contrail-charts/components/navigation/NavigationConfigModel.js
@@ -6,7 +6,7 @@ define([
   'jquery',
   'underscore',
   'd3',
-  'contrail-charts/base/ContrailChartsConfigModel'
+  'contrail-charts/contrail/ContrailChartsConfigModel'
 ], function ($, _, d3, ContrailChartsConfigModel) {
   var NavigationConfigModel = ContrailChartsConfigModel.extend({
     defaults: {
