+++ src/js/contrail-charts/components/tooltip/TooltipConfigModel.js
@@ -5,7 +5,7 @@
 define([
   'jquery',
   'underscore',
-  'contrail-charts/base/ContrailChartsConfigModel'
+  'contrail-charts/contrail/ContrailChartsConfigModel'
 ], function ($, _, ContrailChartsConfigModel) {
   var TooltipConfigModel = ContrailChartsConfigModel.extend({
     defaults: {
