+++ src/js/contrail-charts/contrail/ContrailEvents.js
@@ -1,0 +1,5 @@
+define([
+  'contrail-charts/contrail/plugins/backbone/Events'
+], function (Events) {
+  return Events
+})
