+++ src/js/contrail-charts/contrail/ContrailModel.js
@@ -1,0 +1,5 @@
+define([
+  'contrail-charts/contrail/plugins/backbone/Model'
+], function (Model) {
+  return Model
+})
