+++ src/js/contrail-charts/contrail/ContrailView.js
@@ -1,0 +1,5 @@
+define([
+  'contrail-charts/contrail/plugins/backbone/View'
+], function (View) {
+  return View
+})
